refactor: migrated the last table endpoint
