+++ sites/geohub/src/routes/(app)/api/+page.ts
@@ -4,7 +4,6 @@ export const load: PageLoad = async () => {
 	const title = 'API Docs | GeoHub';
 	const content = 'GeoHub API Spec';
 	const spec = '/api/openapi.json';
-	// const spec = '/api/swagger/spec.json';
 
 	return {
 		title,
